+++ R/is_empty.R
@@ -64,7 +64,8 @@ is_empty <- function(x, first.only = TRUE) {
       })
       # return result for multiple elements of character vector
       if (first.only) {
-        return(unname(zero_len)[1])
+        zero_len <- isTRUE(zero_len[1])
+        if (length(x) > 0) x <- x[1]
       } else {
         return(unname(zero_len))
       }